+++ changelog.md
@@ -8,6 +8,7 @@ Codion Change Log
 ### is.codion.swing.framework.ui
 - EntityTablePanel.SHOW_REFRESH_PROGRESS_BAR renamed REFRESH_PROGRESS_BAR, now true by default.
 - EntitySearchField bug fixed, no longer updates the background color when disabled.
+- EntitySearchField.searchOnFocusLost now final, searchOnFocusLost() removed.
 
 ## 0.18.25
 ### is.codion.common.db